refactor: refactor worker and task metrics to use int64 types
- Change `BusyWorkers` type from `uint64` to `int64`
- Add `CompletedTasks` method to `Metric` interface and implementation
- Update `IncBusyWorker` and `DecBusyWorker` to use `atomic.AddInt64`
- Modify test assertions to use `uint64` and `int64` types
- Change `defaultWorkerCount` and related functions from `int` to `int64`
- Update `NewPool` and related functions to use `int64` for size and worker count
- Add `CompletedTasks` method to `Queue` and update related methods to use `uint64` and `int64` types

+++ metric_test.go
@@ -44,8 +44,8 @@ func TestMetricData(t *testing.T) {
 	}))
 	q.Start()
 	time.Sleep(50 * time.Millisecond)
-	assert.Equal(t, 4, q.SubmittedTasks())
-	assert.Equal(t, 2, q.SuccessTasks())
-	assert.Equal(t, 2, q.FailureTasks())
+	assert.Equal(t, uint64(4), q.SubmittedTasks())
+	assert.Equal(t, uint64(2), q.SuccessTasks())
+	assert.Equal(t, uint64(2), q.FailureTasks())
 	q.Release()
 }
